add loading view and refresh buttons
