+++ CryptoAPI/CryptoAPI/Models/Dado.swift
