+++ CryptoAPI/CryptoAPI/ViewModels/ContentViewModel.swift
@@ -1,0 +1,13 @@
+//
+//  ContentViewModel.swift
+//  CryptoAPI
+//
+//  Created by Guilherme Ferreira Lenzolari on 04/07/24.
+//
+
+import Foundation
+
+@Observable
+class ContentViewModel {
+    
+}
